backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# HTTP Client
httpx==0.28.1

# SSE / Streaming
sse-starlette==2.2.1

# UUID
uuid6==2024.7.10

# Utilities
python-dateutil==2.9.0.post0

# Resilience
tenacity>=8.2.0

# ── Document Automation ─────────────────────────────────
# PDF: HTML+CSS → PDF (recommended for complex layouts)
weasyprint==63.1
# Excel: native .xlsx with charts & formatting
openpyxl==3.1.5
# Word: .docx generation with styles
python-docx==1.1.2
# PPTX: presentation generation
python-pptx==1.0.2
# Templates
jinja2==3.1.4
